Creación componente usuario

+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/controller/EstudioController.java
@@ -55,7 +55,7 @@ public class EstudioController {
     
     @DeleteMapping ("/delete/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity <?> borrarEstudio (@PathVariable("id") Long id){
+    public ResponseEntity <?> borrarEstudio (@PathVariable("id") Integer id){
         iestudioService.eliminarEstudio(id);
         return new ResponseEntity<>(HttpStatus.OK);
