+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/repository/EstudioRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstudioRepository extends JpaRepository<Estudio, Long>{
+public interface EstudioRepository extends JpaRepository<Estudio, Integer>{
     
     
 }
